+++ app/src/main/java/net/ballmerlabs/scatterbrain/network/BlockDataPacket.java
@@ -125,7 +125,7 @@ public class BlockDataPacket extends ScatterStanza {
         //verify crc with stored copy 
         CRC32 crc = new CRC32();
         crc.update(contents,0, contents.length - 4);
-        byte[] check = ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putLong(crc.getValue()).array();
+        byte[] check = ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putInt((int)crc.getValue()).array();
         byte[] real = new byte[4];
         for(int x=0;x<real.length;x++) {
             real[x] = contents[(contents.length - 4) + x];
@@ -178,7 +178,7 @@ public class BlockDataPacket extends ScatterStanza {
         //basic crc for integrity check
         CRC32 crc = new CRC32();
         crc.update(contents,0, contents.length - 4);
-        byte[] c = ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putLong(crc.getValue()).array();
+        byte[] c = ByteBuffer.allocate(4).order(ByteOrder.LITTLE_ENDIAN).putInt((int)crc.getValue()).array();
         for(int x=0;x<c.length;x++) {
             contents[(contents.length - 4) + x] = c[x];
         }
